+++ changelog.md
@@ -1,3 +1,10 @@
+# SDK v.2.5.8
+- Add contact rendering
+- Add location rendering only if `google_key` is passed as options when initiating sdk
+- Modify how reply rendered
+- Fix login-error callback not being called
+- change clickable area for `headerClickCallback`
+
 # SDK v.2.5.7
 - Add `carousel` component
 - Activate sync on loginsuccess